refactor: 重构account项目的服务类、仓库类，抽取相同的方法通过泛型实现抽象类、方法，从而简化实体类仓库、服务的创建

+++ account/src/main/java/org/ziranziyuanting/account/entity/SysOrg.java
@@ -33,11 +33,11 @@ public class SysOrg extends CommonEntity {
 
     /** 扩展信息 */
     private String extJson;
-    public static SysOrg of(int sortCode, Integer parentId, String name, String category){
+    public static SysOrg of(int sortCode, Long parentId, String name, String category){
         SysOrg r = new SysOrg();
         r.setCategory(category);
         r.setSortCode(sortCode);
-        r.setParentId(parentId.longValue());
+        r.setParentId(parentId);
         r.setName(name);
         return r;
     }
